Revert "pmt-ptw"
This reverts commit 8f59a0bf2accbbde5a0b0d1fd2972b73b6595330.

+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/Counter.java
@@ -56,7 +56,7 @@ public final class Counter implements Externalizable
 
     public int getCount()
     {
-        return ((Integer)(this.count)).intValue();
+        return this.count;
     }
 
     public void reset()
