+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/block/procedure/primitive/InjectIntoIntProcedure.java
@@ -34,6 +34,6 @@ public class InjectIntoIntProcedure<T> implements Procedure<T>
 
     public int getResult()
     {
-        return ((Integer)(this.result)).intValue();
+        return this.result;
     }
 }
