+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/block/procedure/primitive/InjectIntoLongProcedure.java
@@ -34,6 +34,6 @@ public class InjectIntoLongProcedure<T> implements Procedure<T>
 
     public long getResult()
     {
-        return ((Long)(this.result)).longValue();
+        return this.result;
     }
 }
